+++ dev_web_react_trabfinal/src/app/Loja.css
@@ -1,52 +1,4 @@
 /* ----------------------------------------- LOJA ---------------------------------------------------- */
-
-.sec {
-    min-height: 100vh;
-    width: 100%;
-    background-image: linear-gradient(rgba(28, 29, 29, 0.7), rgba(61, 66, 78, 0.7)), url(https://images.pexels.com/photos/1103970/pexels-photo-1103970.jpeg);
-    background-position: center;
-    background-size: cover;
-    position: relative;
-  }
-  nav img {
-    width: 75px;
-  }
-  
-  .blur {
-    position: fixed;
-    top: 0;
-    width: 100%;
-    min-height: 90px;
-    z-index:-1;
-    -webkit-backdrop-filter: blur(8px);
-    backdrop-filter: blur(8px);
-  }
-  
-  .nav-links {
-    flex: 1;
-    text-align: right;
-  }
-  
-  .nav-links ul li {
-    list-style: none;
-    display: inline-block;
-    padding: 8px 12px;
-    position: relative;
-  }
-  
-  .nav-links ul li::after {
-    content: '';
-    width: 0%;
-    height: 2px;
-    background: #a4cad4;
-    display: block;
-    margin: auto;
-    transition: 0.5s;
-  }
-  
-  .nav-links ul li:hover::after {
-    width: 100%;
-  }
   
   .loja {
     width: 90%;
