+++ dev_web_react_trabfinal/src/app/Loja.js
@@ -7,6 +7,7 @@ import Reparacao from './Reparacao';
 import Funcionario from './Funcionario';
 import Dispositivo from './Dispositivo';
 import Login from './Login';
+import Navbar from './Navbar';
 
 class Loja extends Component {
     // variável state com as variaveis necessarias
@@ -136,15 +137,6 @@ class Loja extends Component {
             LogId: id
         });
     };
-
-    /*---------------------------------- Estética da App ----------------------------------*/
-
-    scrollToFuncionario = () => {
-        const funcionarioSection = document.getElementById('funcionario');
-        if (funcionarioSection) {
-          funcionarioSection.scrollIntoView({ behavior: 'smooth' });
-        }
-    };
      
     /*------------ Render ------------*/
 
@@ -153,21 +145,7 @@ class Loja extends Component {
         
         return (
             <>
-                <div className='sec'>
-                    <nav className='navbar fixed-top navbar-dark'>
-                        <a class="navbar-brand" href=''><img src='https://cdn.icon-icons.com/icons2/2620/PNG/512/among_us_player_red_icon_156942.png' class="d-inline-block"></img>Amogus</a>
-                        <div className='nav-links'>
-                            <ul>
-                                <li class="nav-item"><a class="nav-link"  data-bs-toggle='offcanvas' data-bs-target='#sidebar' aria-controls='sidebar'>&#9776; HOME</a></li>
-                                <li class="nav-item"><a class="nav-link" onClick={this.scrollToFuncionario}>Funcionarios</a></li>
-                                <li class="nav-item"><a class="nav-link" href=''>Course</a></li>
-                                <li class="nav-item"><a class="nav-link" href=''>IDK</a></li>
-                                <li class="nav-item"><a class="nav-link" href=''>HOE</a></li>
-                            </ul>
-                        </div>
-                        <div className='blur'></div> 
-                    </nav>
-                </div>
+                <Navbar />
 
                 <Login buscar={() => { this.buscarDadosClientes() }}
                     LoginStatus={(logged, id) => {this.handelLoginStatus(logged, id)}}
